refactor(workflow-tasks): move bootstrap component into library, drop ToolPaths coupling
